+++ lib/mdxish.ts
@@ -1,5 +1,6 @@
 import type { CustomComponents } from '../types';
 import type { Root } from 'hast';
+import type { Root as MdastRoot } from 'mdast';
 
 import { mdxExpressionFromMarkdown } from 'mdast-util-mdx-expression';
 import { mdxExpression } from 'micromark-extension-mdx-expression';
@@ -9,9 +10,11 @@ import remarkFrontmatter from 'remark-frontmatter';
 import remarkGfm from 'remark-gfm';
 import remarkParse from 'remark-parse';
 import remarkRehype from 'remark-rehype';
+import remarkStringify from 'remark-stringify';
 import { unified } from 'unified';
 import { VFile } from 'vfile';
 
+import compilers from '../processor/compile';
 import { rehypeMdxishComponents } from '../processor/plugin/mdxish-components';
 import { mdxComponentHandlers } from '../processor/plugin/mdxish-handlers';
 import calloutTransformer from '../processor/transform/callouts';
@@ -23,11 +26,13 @@ import evaluateExpressions from '../processor/transform/mdxish/evaluate-expressi
 import mdxishComponentBlocks from '../processor/transform/mdxish/mdxish-component-blocks';
 import mdxishHtmlBlocks from '../processor/transform/mdxish/mdxish-html-blocks';
 import magicBlockRestorer from '../processor/transform/mdxish/mdxish-magic-blocks';
+import mdxishMermaidTransformer from '../processor/transform/mdxish/mdxish-mermaid';
 import { processSnakeCaseComponent } from '../processor/transform/mdxish/mdxish-snake-case-components';
 import mdxishTables from '../processor/transform/mdxish/mdxish-tables';
 import normalizeEmphasisAST from '../processor/transform/mdxish/normalize-malformed-md-syntax';
 import { preprocessJSXExpressions, type JSXContext } from '../processor/transform/mdxish/preprocess-jsx-expressions';
 import restoreSnakeCaseComponentNames from '../processor/transform/mdxish/restore-snake-case-component-name';
+import { preserveBooleanProperties, restoreBooleanProperties } from '../processor/transform/mdxish/retain-boolean-attributes';
 import variablesTextTransformer from '../processor/transform/mdxish/variables-text';
 import tailwindTransformer from '../processor/transform/tailwind';
 
@@ -43,13 +48,7 @@ export interface MdxishOpts {
 
 const defaultTransformers = [codeTabsTransformer, gemojiTransformer, embedTransformer];
 
-/**
- * Process markdown content with MDX syntax support.
- * Detects and renders custom component tags from the components hash.
- *
- * @see {@link https://github.com/readmeio/rmdx/blob/main/docs/mdxish-flow.md}
- */
-export function mdxish(mdContent: string, opts: MdxishOpts = {}): Root {
+export function mdxishAstProcessor(mdContent: string, opts: MdxishOpts = {}) {
   const { components: userComponents = {}, jsxContext = {}, useTailwind, format } = opts;
 
   const components: CustomComponents = {
@@ -90,9 +89,52 @@ export function mdxish(mdContent: string, opts: MdxishOpts = {}): Root {
     .use(evaluateExpressions, { context: jsxContext }) // Evaluate MDX expressions using jsxContext
     .use(variablesTextTransformer) // Parse {user.*} patterns from text (can't rely on remarkMdx)
     .use(useTailwind ? tailwindTransformer : undefined, { components: tempComponentsMap })
-    .use(remarkGfm)
+    .use(remarkGfm);
+
+  return {
+    processor,
+    /**
+     * @todo we need to return this transformed content for now
+     * but ultimately need to properly tokenize our special markdown syntax
+     * into hast nodes instead of relying on transformed content
+     */
+    parserReadyContent,
+  };
+}
+
+/**
+ * Converts an Mdast to a Markdown string.
+ */
+export function mdxishMdastToMd(mdast: MdastRoot) {
+  const md = unified().use(remarkGfm).use(compilers).use(remarkStringify, {
+    bullet: '-',
+    emphasis: '_',
+  }).stringify(mdast);
+  return md;
+}
+
+/**
+ * Processes markdown content with MDX syntax support and returns a HAST.
+ * Detects and renders custom component tags from the components hash.
+ *
+ * @see {@link https://github.com/readmeio/rmdx/blob/main/docs/mdxish-flow.md}
+ */
+export function mdxish(mdContent: string, opts: MdxishOpts = {}): Root {
+  const { components: userComponents = {} } = opts;
+
+  const components: CustomComponents = {
+    ...loadComponents(),
+    ...userComponents,
+  };
+
+  const { processor, parserReadyContent } = mdxishAstProcessor(mdContent, opts);
+
+  processor
     .use(remarkRehype, { allowDangerousHtml: true, handlers: mdxComponentHandlers })
+    .use(preserveBooleanProperties) // RehypeRaw converts boolean properties to empty strings
     .use(rehypeRaw, { passThrough: ['html-block'] })
+    .use(restoreBooleanProperties)
+    .use(mdxishMermaidTransformer) // Add mermaid-render className to pre wrappers
     .use(rehypeSlug)
     .use(rehypeMdxishComponents, {
       components,